[CLEAN] LOGIC clean up simplify position updates
