+++ protocol/x/subaccounts/keeper/position_updates.go
@@ -2,265 +2,122 @@ package keeper
 
 import (
 	"fmt"
-	"sort"
 
 	"github.com/StreamFinance-Protocol/stream-chain/protocol/dtypes"
 	"github.com/StreamFinance-Protocol/stream-chain/protocol/x/subaccounts/types"
 )
-
-// getUpdatedAssetPositions filters out all the asset positions on a subaccount that have
-// been updated. This will include any asset postions that were closed due to an update.
-func getUpdatedAssetPositions(
-	update SettledUpdate,
-) []*types.AssetPosition {
-	assetIdToPosition := createMapFromSliceForPositions(update.SettledSubaccount.AssetPositions)
-	updatedAssetIds := createIdsToEmptyStructMap(update.AssetUpdates)
-
-	updatedAssetPositions := createUpdatedPositionsSlice(
-		updatedAssetIds,
-		assetIdToPosition,
-		func(id uint32) *types.AssetPosition {
-			return &types.AssetPosition{
-				AssetId:  id,
-				Quantums: dtypes.ZeroInt(),
-			}
-		},
-	)
-
-	sortByIdAscending(updatedAssetPositions)
-
-	return updatedAssetPositions
-}
-
-// getUpdatedPerpetualPositions filters out all the perpetual positions on a subaccount that have
-// been updated. This will include any perpetual postions that were closed due to an update or that
-// received / paid out funding payments.
-func getUpdatedPerpetualPositions(
-	update SettledUpdate,
-	fundingPayments map[uint32]dtypes.SerializableInt,
-) []*types.PerpetualPosition {
-	perpetualIdToPosition := createMapFromSliceForPositions(update.SettledSubaccount.PerpetualPositions)
-
-	// `updatedPerpetualIds` indicates which perpetuals were either explicitly updated
-	// (through update.PerpetualUpdates) or implicitly updated (had non-zero last funding
-	// payment).
-	updatedPerpetualIds := createIdsToEmptyStructMap(update.PerpetualUpdates)
-
-	// Mark perpetuals with non-zero funding payment also as updated.
-	for perpetualIdWithNonZeroLastFunding := range fundingPayments {
-		updatedPerpetualIds[perpetualIdWithNonZeroLastFunding] = struct{}{}
-	}
-
-	// Properties besides the PerpetualId and Quantums are left as the default values as
-	// a 0-sized position indicates the position is closed and thus the funding index and
-	// the side of the position does not matter.
-	updatedPerpetualPositions := createUpdatedPositionsSlice(
-		updatedPerpetualIds,
-		perpetualIdToPosition,
-		func(id uint32) *types.PerpetualPosition {
-			return &types.PerpetualPosition{
-				PerpetualId: id,
-				Quantums:    dtypes.ZeroInt(),
-			}
-		},
-	)
-
-	sortByIdAscending(updatedPerpetualPositions)
-
-	return updatedPerpetualPositions
-}
 
 func UpdateSubaccountPositions(
 	settledUpdates []SettledUpdate,
 	perpIdToFundingIndex map[uint32]dtypes.SerializableInt,
 	perpIdToYieldIndex map[uint32]string,
 ) {
-	for i, update := range settledUpdates {
-		perpetualPositionsMap := createMapFromSliceForPositions(update.SettledSubaccount.PerpetualPositions)
-
-		assetPositionsMap := createMapFromSliceForPositions(update.SettledSubaccount.AssetPositions)
-
-		updatedPerpetualPositions := createUpdatedPerpetualPositions(update, perpetualPositionsMap, perpIdToFundingIndex, perpIdToYieldIndex)
-		settledUpdates[i].SettledSubaccount.PerpetualPositions = updatedPerpetualPositions
-
-		updatedAssetPositions := createUpdatedAssetPositions(update, assetPositionsMap)
-		settledUpdates[i].SettledSubaccount.AssetPositions = updatedAssetPositions
+	for i := range settledUpdates {
+		update := &settledUpdates[i]
+		updatePerpetualPositions(update, perpIdToFundingIndex, perpIdToYieldIndex)
+		updateAssetPositions(update)
 	}
 }
 
-func createUpdatedPerpetualPositions(
-	update SettledUpdate,
-	perpetualIdToPerpetualPosition map[uint32]*types.PerpetualPosition,
+func updatePerpetualPositions(
+	update *SettledUpdate,
 	perpIdToFundingIndex map[uint32]dtypes.SerializableInt,
 	perpIdToYieldIndex map[uint32]string,
-) []*types.PerpetualPosition {
-	for _, perpetualUpdate := range update.PerpetualUpdates {
-		perpetualPosition, exists := perpetualIdToPerpetualPosition[perpetualUpdate.PerpetualId]
-		if exists {
-			updateExistingPosition(
-				perpetualPosition,
-				perpetualUpdate,
-				perpetualIdToPerpetualPosition,
-				perpetualUpdate.PerpetualId,
-			)
-		} else {
-			insertNewPerpetualPosition(
-				perpetualUpdate,
-				perpetualIdToPerpetualPosition,
-				perpIdToFundingIndex,
-				perpIdToYieldIndex,
-			)
+) {
+	for _, perpUpdate := range update.PerpetualUpdates {
+		if idx := findPosition(update.SettledSubaccount.PerpetualPositions, perpUpdate.PerpetualId); idx >= 0 {
+			updateExistingPerpPosition(update, idx, perpUpdate)
+		} else if perpUpdate.BigQuantumsDelta.Sign() != 0 {
+			createNewPerpPosition(update, perpUpdate, perpIdToFundingIndex, perpIdToYieldIndex)
 		}
 	}
-
-	perpetualPositions := createSliceFromMapForPositions(perpetualIdToPerpetualPosition)
-	sortByIdAscending(perpetualPositions)
-
-	return perpetualPositions
 }
 
-func createUpdatedAssetPositions(
-	update SettledUpdate,
-	assetIdToAssetPosition map[uint32]*types.AssetPosition,
-) []*types.AssetPosition {
+func updateAssetPositions(update *SettledUpdate) {
 	for _, assetUpdate := range update.AssetUpdates {
-		assetPosition, exists := assetIdToAssetPosition[assetUpdate.AssetId]
-		if exists {
-			updateExistingPosition(
-				assetPosition,
-				assetUpdate,
-				assetIdToAssetPosition,
-				assetUpdate.AssetId,
-			)
-		} else {
-			insertNewAssetPosition(
-				assetUpdate,
-				assetIdToAssetPosition,
-			)
+		if idx := findPosition(update.SettledSubaccount.AssetPositions, assetUpdate.AssetId); idx >= 0 {
+			updateExistingAssetPosition(update, idx, assetUpdate)
+		} else if assetUpdate.BigQuantumsDelta.Sign() != 0 {
+			createNewAssetPosition(update, assetUpdate)
 		}
 	}
-
-	assetPositions := createSliceFromMapForPositions(assetIdToAssetPosition)
-	sortByIdAscending(assetPositions)
-
-	return assetPositions
 }
 
-func sortByIdAscending[T types.PositionSize](items []T) {
-	sort.Slice(items, func(i, j int) bool {
-		return items[i].GetId() < items[j].GetId()
-	})
-}
+// Helper functions
 
-// Note: We expect this function to be called with a slice of pointers
-// in order to avoid copying the entire position struct.
-func createMapFromSliceForPositions[P types.PositionSize](
-	positions []P,
-) map[uint32]P {
-	positionsMap := make(map[uint32]P)
-	for _, position := range positions {
-		positionsMap[position.GetId()] = position
-	}
-	return positionsMap
-}
-
-func createSliceFromMapForPositions[P any](
-	idToPosition map[uint32]*P,
-) []*P {
-	positions := make([]*P, 0, len(idToPosition))
-	for _, value := range idToPosition {
-		positions = append(positions, value)
-	}
-	return positions
-}
-
-func createIdsToEmptyStructMap[P types.PositionSize](
-	updates []P,
-) map[uint32]struct{} {
-	ids := make(map[uint32]struct{})
-	for _, update := range updates {
-		ids[update.GetId()] = struct{}{}
-	}
-	return ids
-}
-
-func createUpdatedPositionsSlice[P types.Position](
-	ids map[uint32]struct{},
-	idToPosition map[uint32]P,
-	createEmptyPosition func(uint32) P,
-) []P {
-	updatedPositions := make([]P, 0, len(ids))
-	for id := range ids {
-		position, exists := idToPosition[id]
-
-		// If a position does not exist on the subaccount with the id of an update, it must
-		// have been deleted due to quantums becoming 0. This needs to be included in the event,
-		// so we construct a position with the id of the update and a Quantums value of 0.
-		if !exists {
-			position = createEmptyPosition(id)
+func findPosition[P interface{ GetId() uint32 }](positions []P, id uint32) int {
+	for i, pos := range positions {
+		if pos.GetId() == id {
+			return i
 		}
-		updatedPositions = append(updatedPositions, position)
 	}
-	return updatedPositions
+	return -1
 }
 
-func updateExistingPosition[P types.Position, T types.PositionSize](
-	position P,
-	update T,
-	idToPosition map[uint32]P,
-	id uint32,
-) {
-	curQuantums := position.GetBigQuantums()
-	updateQuantums := update.GetBigQuantums()
-	newQuantums := curQuantums.Add(curQuantums, updateQuantums)
-
-	position.SetBigQuantums(newQuantums)
-
+func updateExistingPerpPosition(update *SettledUpdate, idx int, perpUpdate types.PerpetualUpdate) {
+	pos := update.SettledSubaccount.PerpetualPositions[idx]
+	newQuantums := pos.Quantums.BigInt().Add(
+		pos.Quantums.BigInt(),
+		perpUpdate.BigQuantumsDelta,
+	)
 	if newQuantums.Sign() == 0 {
-		delete(idToPosition, id)
+		update.SettledSubaccount.PerpetualPositions = append(
+			update.SettledSubaccount.PerpetualPositions[:idx],
+			update.SettledSubaccount.PerpetualPositions[idx+1:]...,
+		)
+	} else {
+		pos.Quantums = dtypes.NewIntFromBigInt(newQuantums)
 	}
 }
 
-func insertNewAssetPosition(
-	assetUpdate types.AssetUpdate,
-	assetPositionsMap map[uint32]*types.AssetPosition,
-) {
-	if assetUpdate.GetBigQuantums().Sign() == 0 {
-		return
-	}
-
-	assetPosition := &types.AssetPosition{
-		AssetId:  assetUpdate.AssetId,
-		Quantums: dtypes.NewIntFromBigInt(assetUpdate.GetBigQuantums()),
-	}
-
-	assetPositionsMap[assetUpdate.AssetId] = assetPosition
-}
-
-func insertNewPerpetualPosition(
-	perpetualUpdate types.PerpetualUpdate,
-	perpetualPositionsMap map[uint32]*types.PerpetualPosition,
+func createNewPerpPosition(
+	update *SettledUpdate,
+	perpUpdate types.PerpetualUpdate,
 	perpIdToFundingIndex map[uint32]dtypes.SerializableInt,
 	perpIdToYieldIndex map[uint32]string,
 ) {
-	fundingIndex, exists := perpIdToFundingIndex[perpetualUpdate.PerpetualId]
+	fundingIndex, exists := perpIdToFundingIndex[perpUpdate.PerpetualId]
 	if !exists {
-		// Invariant: `perpIdToFundingIndex` contains all existing perpetauls,
-		// and perpetual position update must refer to an existing perpetual.
-		panic(fmt.Sprintf("perpetual id %d not found in perpIdToFundingIndex", perpetualUpdate.PerpetualId))
+		panic(fmt.Sprintf("perpetual id %d not found in perpIdToFundingIndex", perpUpdate.PerpetualId))
 	}
-
-	yieldIndex, exists := perpIdToYieldIndex[perpetualUpdate.PerpetualId]
+	yieldIndex, exists := perpIdToYieldIndex[perpUpdate.PerpetualId]
 	if !exists {
-		panic(fmt.Sprintf("perpetual id %d not found in perpIdToYieldIndex", perpetualUpdate.PerpetualId))
+		panic(fmt.Sprintf("perpetual id %d not found in perpIdToYieldIndex", perpUpdate.PerpetualId))
 	}
 
-	perpetualPosition := &types.PerpetualPosition{
-		PerpetualId:  perpetualUpdate.PerpetualId,
-		Quantums:     dtypes.NewIntFromBigInt(perpetualUpdate.GetBigQuantums()),
-		FundingIndex: fundingIndex,
-		YieldIndex:   yieldIndex,
-	}
+	update.SettledSubaccount.PerpetualPositions = append(
+		update.SettledSubaccount.PerpetualPositions,
+		&types.PerpetualPosition{
+			PerpetualId:  perpUpdate.PerpetualId,
+			Quantums:     dtypes.NewIntFromBigInt(perpUpdate.BigQuantumsDelta),
+			FundingIndex: fundingIndex,
+			YieldIndex:   yieldIndex,
+		},
+	)
+}
 
-	perpetualPositionsMap[perpetualUpdate.PerpetualId] = perpetualPosition
+func updateExistingAssetPosition(update *SettledUpdate, idx int, assetUpdate types.AssetUpdate) {
+	pos := update.SettledSubaccount.AssetPositions[idx]
+	newQuantums := pos.Quantums.BigInt().Add(
+		pos.Quantums.BigInt(),
+		assetUpdate.BigQuantumsDelta,
+	)
+	if newQuantums.Sign() == 0 {
+		update.SettledSubaccount.AssetPositions = append(
+			update.SettledSubaccount.AssetPositions[:idx],
+			update.SettledSubaccount.AssetPositions[idx+1:]...,
+		)
+	} else {
+		pos.Quantums = dtypes.NewIntFromBigInt(newQuantums)
+	}
+}
+
+func createNewAssetPosition(update *SettledUpdate, assetUpdate types.AssetUpdate) {
+	update.SettledSubaccount.AssetPositions = append(
+		update.SettledSubaccount.AssetPositions,
+		&types.AssetPosition{
+			AssetId:  assetUpdate.AssetId,
+			Quantums: dtypes.NewIntFromBigInt(assetUpdate.BigQuantumsDelta),
+		},
+	)
 }
